refactor: MoHeader 통합
